refactor big main class into smaller mixins
